requirements.txt
flask
pandas
psycopg2-binary
python-dotenv
coverage
scikit-learn
numpy
